% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_fn.R
\name{frac_diff}
\alias{frac_diff}
\title{frac_diff}
\usage{
frac_diff(y, d)
}
\arguments{
\item{y}{A vector time series}

\item{d}{Order of differencing}
}
\value{
The d-th difference of y
}
\description{
Takes the fractional difference of a vector time series
}
