% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_fn.R
\name{frac_diff_multi}
\alias{frac_diff_multi}
\title{frac_diff_multi}
\usage{
frac_diff_multi(y, d)
}
\arguments{
\item{y}{A (T x N) matrix of N time series}

\item{d}{Order of differencing}
}
\value{
The d-th difference of y
}
\description{
Takes the fractional difference of a matrix of inputs (column-wise)
}
