% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fUC.R
\name{fUC_comp}
\alias{fUC_comp}
\title{fUC_comp}
\usage{
fUC_comp(y, d, nu, ar, corr = FALSE)
}
\arguments{
\item{y}{Series to be decomposed into trend and cycle}

\item{d}{Memory parameter (integration order of trend)}

\item{nu}{Either a scalar giving the variance-ratio of cycle shock and trend shock
(if corr = FALSE), or the 2x2 covariance matrix of trend and cycle innovations Q.}

\item{ar}{Autoregressive coefficients of the cycle}

\item{corr}{Either TRUE or FALSE. If TRUE, trend and cycle are allowed to be correlated,
and nu must specify the 2x2 covariance matrix. If FALSE, trend and cycle are uncorrelated,
and nu can be a scalar.}
}
\value{
A list with elements x, the estimated trend component, c, the estimated cycle component,
and v, the one-step ahead prediction error
}
\description{
Kalman filter based calculation of fractional trend and cycle. The
function evaluates the analytical solution of the Kalman filter for fractional
trend and cycle given some parameters and a series to be decomposed
}
