% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fUC.R
\name{fUC_opt}
\alias{fUC_opt}
\title{fUC_opt}
\usage{
fUC_opt(
  theta,
  y,
  nulim = c(0.05, 10),
  quiet = TRUE,
  deterministics = FALSE,
  corr = FALSE,
  START = 1,
  return.det = FALSE,
  d.int = c(0, 2),
  eta = NULL
)
}
\arguments{
\item{theta}{Parameter vector of form (d, mlogvech(Q), ar)}

\item{y}{Series to be decomposed into trend and cycle}

\item{nulim}{Interval for nu}

\item{quiet}{If FALSE, the value of the objective function is printed out}

\item{deterministics}{Either FALSE, or "const" for constant, "trend" for linear trend,
or "frac" for trend of order d}

\item{corr}{Either TRUE or FALSE. If TRUE, trend and cycle are allowed to be correlated,
and nu must specify the 2x2 covariance matrix. If FALSE, trend and cycle are uncorrelated,
and nu can be a scalar.}

\item{START}{First observable y that enters the objective function}

\item{return.det}{If true, estimates for the deterministic components are returned}

\item{d.int}{Interval for integration order#}
}
\value{
Sum of squared prediction errors for a given parameter vector theta and an
observable series y
}
\description{
Objective function for CSS estimator of fractional UC model
}
