% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fUC.R
\name{fUC_opt_ML}
\alias{fUC_opt_ML}
\title{fUC_opt_ML}
\usage{
fUC_opt_ML(
  theta,
  y,
  nulim = c(0.05, 10),
  quiet = TRUE,
  approx = TRUE,
  corr = FALSE,
  deterministics = FALSE,
  START = 1,
  diffuse = FALSE,
  return.det = FALSE,
  nu.opt = FALSE,
  d.int = c(0, 2),
  eta = NULL,
  Q.trans = "mlv",
  flip = F,
  neg = F
)
}
\arguments{
\item{theta}{Parameter vector of form (d, mlogvech(Q), ar)}

\item{y}{Series to be decomposed into trend and cycle}

\item{nulim}{Interval for nu}

\item{quiet}{If FALSE, the value of the objective function is printed out}

\item{approx}{If TRUE, the steady state Kalman filter will be used once the variance of
the prediction error has sufficiently converged}

\item{corr}{Either TRUE or FALSE. If TRUE, trend and cycle are allowed to be correlated,
and nu must specify the 2x2 covariance matrix. If FALSE, trend and cycle are uncorrelated,
and nu can be a scalar.}

\item{deterministics}{Either FALSE, or "const" for constant, "trend" for linear trend,
or "frac" for trend of order d}

\item{START}{First observable y that enters the objective function}

\item{diffuse}{If TRUE, the Kalman filter will be initialized diffusely}

\item{return.det}{If true, estimates for the deterministic components are returned}

\item{nu.opt}{If TRUE, optimization will be conducted setting the variance of
the long run innovations to unity. Highly recommended to set this to FALSE, unless
the QML estimator is compared to the CSS estimator in a simulation}

\item{d.int}{Interval for integration order#}

\item{Q.trans}{Type of transformation for covariance matrix in the estimation. Use
of "mlv" is highly recommended}
}
\value{
Negative log likelihood for a given parameter vector theta and an
observable series y
}
\description{
Objective function for QML estimator of fractional UC model
}
