% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_fn.R
\name{mlogvech2mat}
\alias{mlogvech2mat}
\title{mlogvech2mat}
\usage{
mlogvech2mat(x, eta = NULL)
}
\arguments{
\item{x}{Log-transformed matrix}
}
\value{
Inverse log-transformed matrix
}
\description{
Matrix inverse log-transform
}
