% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_fn.R
\name{pd_cov}
\alias{pd_cov}
\title{pd_cov}
\usage{
pd_cov(theta, type = "mlv", flip = F, neg = F)
}
\arguments{
\item{theta}{Transformed parameters to be optimized over}

\item{type}{Type of transformation. Possible types are "mlv" (matrix exponential),
"chol" (cholesky transformation), pchol (transformewd cholesky transformation),
and "beta" (beta transformation). "mlv" is suggested}
}
\value{
p.d. matrix
}
\description{
Different matrix transformations that ensure p.d.
}
